requirements.txt
requests >= 2.20.0
beautifulsoup4 >= 4.4.0
pyyaml >= 3.13
jinja2 >= 2.10
